# Conflicts:
#	app/www/app.html
#	app/www/index.html
#	src/app/main/xhsc/ys/evaluateQuesDetail.html
#	src/app/main/xhsc/ys/evaluatelist.html

+++ src/app/main/xhsc/ys/evaluatelistPcController.js
@@ -350,7 +350,7 @@
         }
       ).then(function (result) {
         if(result){
-          var r = parseInt(result.value);
+          var r = parseFloat(result.value);
           if(!isNaN(result.value)){
             if(item.Weight<result.value || result.value<0){
               utils.alert('输入的值应该介于0 与 '+item.Weight + ' 之间');
